refactor empty state into its own element
