+++ pages/main.js
@@ -1,73 +1,12 @@
 'use strict'
 
 const html = require('choo/html')
-const css = require('sheetify')
 
 const Header = require('../elements/header')
-const button = require('../elements/button')
 const sprite = require('../elements/sprite')
 const Table = require('../elements/table')
-const icon = require('../elements/icon')
 const Welcome = require('../elements/welcome')
-
-const skeleton = css`
-  :host {
-    position: relative;
-    .skeleton {
-      position: fixed;
-      top: 3.5rem;
-      left: 1.25rem;
-      width: 232px;
-      max-width: 100vw;
-    }
-    .dotted-lines {
-      position: absolute;
-      top: .25rem;
-      right: 5.5rem;
-      width: 17rem;
-      z-index: 3;
-    }
-    .create-new-dat,
-    .link {
-      position: absolute;
-      width: 15rem;
-    }
-    .create-new-dat {
-      top: 14.5rem;
-      right: 4rem;
-      svg {
-        display: inline-block;
-        width: 2rem;
-        height: 2rem;
-      }
-    }
-    .link {
-      top: 6rem;
-      right: 8.5rem;
-      color: red;
-      svg {
-        display: inline-block;
-        width: 2rem;
-        height: 2rem;
-        margin-bottom: -.75rem;
-      }
-    }
-  }
-`
-
-const welcome = css`
-  :host {
-    height: 100vh;
-    background-color: var(--color-neutral);
-    color: var(--color-white);
-    display: flex;
-    flex-direction: column;
-    align-items: center;
-    justify-content: center;
-    text-align: center;
-    -webkit-app-region: drag;
-  }
-`
+const Empty = require('../elements/empty')
 
 module.exports = mainView
 
@@ -109,7 +48,7 @@ function mainView (state, emit) {
       <div>
         ${sprite()}
         ${header}
-        ${EmptyState()}
+        ${Empty()}
       </div>
     `
   }
@@ -129,35 +68,4 @@ function mainView (state, emit) {
       emit('hide welcome screen')
     }
   }
-}
-
-function EmptyState () {
-  return html`
-    <main class="${skeleton}">
-      <img src="./public/img/table-skeleton.svg" alt="" class="skeleton">
-      <div class="tutorial">
-        <img src="./public/img/dotted-lines.svg" alt="" class="dotted-lines">
-        <div class="link">
-          ${icon('link', { class: 'color-blue-disabled' })}
-          <h3 class="f4 ttu mt0 mb0 color-blue-disabled">
-            Import Dat
-          </h3>
-          <p class="f7 color-neutral-40">
-            Download an existing dataset
-            <br>
-            by entering its dat link…
-          </p>
-        </div>
-        <div class="tr create-new-dat">
-          ${icon('create-new-dat', { class: 'color-green-disabled' })}
-          <h3 class="f4 ttu mt0 mb0 color-green-disabled">Create New Dat</h3>
-          <p class="f7 color-neutral-40">
-            … or select one of your local
-            <br>
-            datasets and start sharing it.
-          </p>
-        </div>
-      </div>
-    </main>
-  `
 }
